preparations for daemonizing:     - factor most of main() into dedicated functions / TUs     - allow for multiple runs a la multiple configs (simple loop for now)     - close per-run memory leaks a la valgrind
