+++ src/Digitizer.cc
@@ -42,4 +42,8 @@ size_t Digitizer::readoutBLT(char *buffer, size_t buffer_size) {
     return offset;
 }
 
+Decoder::~Decoder(){
+    /**/
+}
+
 void Decoder::dispatch(int nfd, int *fds) { }
